requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.33
pandas==1.5.3
